requirements.txt
Flask==3.1.0
pandas==2.2.3
plotly==5.24.1
statsmodels==0.14.6
numpy==2.1.3
google-genai
python-dotenv
markdown
